add fastrpc
FastRPC is a communication protocol used to communicate with some remote
processors on Qualcomm SoC's. It might be needed to configure sensors
and to prevent an ADSP crash on the Pixel 3a.

The programs provided by Google and Linaro don't work out of the box to
set up an ADSP listener (even with kernel patches), and is a bit too
unpleasant to read for my tastes. Add a new FastRPC implementation that
will hopefully be more readable for everyone.