fastrpc: treat returned length as a normal number (BREAKING CHANGE)
A FastRPC function call does not necessarily return the amount of bytes
written in the output buffer. Implementing a field that makes these
sizes optional would require the function to keep track of the returned
sizes separately from any other output from this function. Since
returned lengths are interleaved and at the last part of the argument
list, it already needs an unnecessary pass to write back the lengths.
Move the returned lengths to the "returned numbers" part of the argument
list and assume that they are counted in the out_nums property of the
function definition.

The (maximum) sizes of the buffers should still be handled separately
because they need to be passed both to the kernel and to the remote
processor. Removing this special handling would just duplicate the size
arguments, while its implementation is more maintainable than the
returned lengths. The kernel API would need to change for this separate
handling to be unnecessary. Therefore, it is useless to introduce any
other breaking change while old code is already broken.