fastrpc: write to correct first input buffer index
The first input buffer is skipped when passing the function-level input
buffers, but the first entry communicating the buffer's size is also
skipped. This is incorrect: the sizes for these buffers start at
inbuf[def->in_nums].  Start the iterator at zero and add one to the
argument index to make it clear that the argument buffers index is
offset and the first buffer index is not.

Fixes: 9adc9ea326d6a ("add fastrpc")