fastrpc: only skip the first output buffer if it exists
The function-level output buffers are passed with the assumption that
a first output buffer needs to be skipped. This is always true, given
that there is always a returned length of each buffer, but it is not
always the case. Calculate the offset instead of hardcoding it so the
first fixed-length output buffer is only skipped if it exists.