fastrpc: only populate the first arguments if needed
If there are no 32-bit integers in the first input or output buffers,
then the argument pointer might be NULL or point out-of-bounds. Only
write to this pointer's target if the buffer was allocated, otherwise
the size of the corresponding first argument has a length of zero and
should be omitted.