fastrpc: listener: pass correct arguments to adsp_listener_next2()
The scalars argument is an output and its initialization does not have
any effect. Instead, the result is set to the highest possible 32-bit
unsigned integer value. The context should always be set to zero. Pass
the correct arguments to the procedure call as specified in
fastrpc.git/src/listener_android.c.