fastrpc: adsp_listener: properly export definitions
The function definitions used to call fastrpc() are silly. They expect
to only be used in one compilation unit, otherwise there is a linker
error. The definitions are not marked as constant, even though changing
the definition at runtime is discouraged and has no practical use.
Properly export the function definitions for the interface.