fastrpc: iobuffer: set buffer types to void *; const when appropriate
The decoder expects a raw buffer that can be written to, even though
there is no limitation that requires this. Character pointers are
convenient for indexing, but inconvenient for casting to different
objects (like parameter sets of 32-bit named integers). Accept generic
pointers to read-only memory as raw buffers and define the pointers in
the I/O buffers as generic.