fastrpc: add hexagonfs
The DSP demands a file system, but shouldn't have unrestricted access to
it (especially if the reverse tunnel is running as root). It makes
requests to files with no regard to any Linux FHS and a sysfs file that
doesn't exist on mainline. Add a virtual file system to take care of
the DSP's needs.