fastrpc: hexagonrpcd: apps_std: do not fail without unrelated dirs
The open_dirs() function opens common directories that are usually
requested multiple times by the remote processor. When it fails to open
a directory, the failure propagates to the remote processor. Since the
remote processor might not need every directory, silence this failure so
the daemon can serve files unrelated to any missing or inaccessible
directories.