tools: registry: store option character as integer
The option character is being compared to -1, but this -1 is an integer.
With GCC 12.2.1 on aarch64 platforms, the 8-bit character is not
properly sign-extended to 32 bits. Store the option character as an
integer so it is properly sign-extended.

Further investigation is needed to determine whether this is an
optimization bug and what value the 8-bit signed -1 is equal to.