fastrpc: hexagonrpcd: iobuffer: only align at beginning of inbuf
When the decoding continues from a misaligned part of the buffer, the
decoder attempts to skip some of the buffer to realign itself, even if
it is not the beginning of the buffer. To allow the decoding of an input
with a size indivisible by 8, only consume the alignment at the
beginning of an input buffer.

The bug outlined here does not happen in practice because the size of
the encoded input is hardcoded to 256 in listener.c.

Fixes: 77f855ec84d9 ("fastrpc: decode input buffers")