fastrpc: hexagonrpcd: iobuffer: change types in decoder context
There is no real reason for the position attributes to be signed. It
does not make sense to have a negative offset in the current input
buffer, negative alignment, or negative offset in the size of the
current input buffer. A negative number of input buffers also makes no
sense. Reorganize the decoder context struct and the I/O buffer struct
so offsets and maximums are together, nothing has the unnecessary
capability of being negative, and the sizes are not wider than possible
in the format.

Also include the definition for uint32_t.

Also update the error message format in listener.c to read a 32-bit
integer.